Add Media Source Engine Step in Load Graph
This change takes the part of |onLoad| that creates media source
engine and makes it into its own state in the graph.

This allows us to pre-initialize media source engine and will allow for
the load-process to be aborted after initializing media source engine
but before load.

This ensure that we only initialize media source engine once per load
(fixing #1570) by modeling "initialized media source" as its own
node in the graph.

The tests that appear to be removed from "player_integration.js" have
been moved to be with the other load graph tests and re-written to
make use of the load-graph events for testing.

The stats test that appears to be removed from "player_unit.js" was
moved to "player_integration.js". This test was failing before the
player was not overriding media source (like the other tests were) since
this one had to do with load order, it was moved to be closer to the
other load order tests.

Issue #816
Issue #997
Closes #1570

Change-Id: I646f0559e3878f28374a3fc56f3d70f6d8a462a4