Refactor Manifest Parser Creation
Rather than have the manifest parser class return the factory
that we then need to call "new" on, we can just have the manifest
parser class return the parser.

This makes the call-sites in Player and Storage simpler and moves
the use of factory+new to one spot so that it will be easier to
remove in 2.6.

Change-Id: I2abb8da603c9226c230bb761a20eb8dd20fc8455