Add Parse Manifest Step in Load Graph
This change takes the part of |onLoad| that created the manifest parser
and parses the manifest and makes it into its own state in the graph.

This change tried to unravel some of the code in play, which causes
some changes to how we created manifest parsers. Since we changed how
we created manifest parsers, that naturally leads to |Storage| needing
a small change since it is the only other place where parsers are
created.

Issue #816
Issue #997

Change-Id: I905a40a1146aed3ef03aba51e33051fef0b79281