Fix HLS not waiting for updates to finish
In all manifest parsers, calls to |update| should resolve when
they are done updating. However the HLS manifest parser did not
return the promise when updating each media playlist. This meant
that the call to |update| would resolve before the update was
complete.

To help ensure that the promise is returned, the method was
converted to use async-await which appears to be easier for the
compiler to catch problems with.

Issue #1157
Issue #1788

Change-Id: I2cfda45c849ea5b4e87d7ff068319f2bf1fcfaa9