Fix cast button on ui.
Previously, the cast button would not show up on the ui.
This was due to a typo in an event listener (it was looking for
'caststatuschange' rather than 'caststatuschanged'), and due to
the ui starting the button out hidden and only revealing it on
cast status change.
This fixes that typo, and changes the button to start out in an
appropriate state.

Issue #1803

Change-Id: I06a18015511fd897b15606d2bb75b3a448e124c7