Add TODO for Walker Error Handling
Right now we have a very crude approach to handling errors which will
not work well with aborting networking retries in the manifest parser.

Instead we want to return to attached over detached whenever possible.
While the change in code is simple, we need to have the tests to verify
this behaviour.

For now, this adds the TODO to document the change that should come in
the near future.

Change-Id: I97e95eade0e23de4b5c0e1abb05e240cb7e6b7bb