Split Manifest Load Step
This CL splits the manifest step into a parser and a manifest step. This
ensures that the abortable operation for the manifest step will always
abort.

Before, it would be possible to abort before the parser was created. If
this happened, the step would not abort, and would run to completion.

Issue #816
Issue #997

Change-Id: If10fe2f618e9352c8a8eb4da2a20879856686c0e