Define HLS Parser Interruption Test
If we try to stop the HLS parser while it is parsing a manifest (either
via |start| or |update|) it can misinterpret the aborted partial segment
request as the server not supporting the range request.

This change creates a test that show that we can't interrupt the HLS
parser while it is parsing the manifest (even through we expect to).

The test added in this CL is expected to fail, showing the problem, and
will be fixed in a following CL, showing the fix works.

Issue #1788

Change-Id: I91c459eb5fe0d84af8fe5987a459a054d3aa1107