Fix HLS parser treating abort as no feature support
When the HLS Parser is shutting down and aborting its in-progress networking
requests, it will misinterpret the aborted operation as "Unable to fetch a
partial HLS segment! Falling back to a full segment request, which is
expensive!"

The short term fix for this is to check the error and exit if it was an
operation aborted error. A TODO was added to point toward a more robust
long term fix.

Issue #1788

Change-Id: I28ecfb9a087eb9fe9cafad0a87032685a92d1be7