Have Walker Surface Error
When the walker encounters an error, it handles the error but it did not
surface the error. This made it hard to log the error without putting
the log statement in the walker.

This changes the walker to pass the error to the handle error method so
that the class using the player (e.g. Player) can use the error in
determining how to respond to the error.

Change-Id: Ied91f8fcf80d5dd8058e8ba58de16d25213c2f0b