Limit Player.createDrmEngine responsibilities
We have a method on player to create a drm engine instance. This method
allows us to inject a fake drm engine during testing.

Before, this method did a lot. It created, configured, and initialized
a drm engine instance. However, in our tests we only created the
instance. For greater consistency between production and test, this CL
moves the configuring and initializing logic into the main load method.

Issue #816
Issue #997

Change-Id: Ie9278b401e89014825c62d1b2c1758b3079f99e9