Correct oversights in the player's load graph
There were some oversights when moving code out of Player.onLoad_. The
method header needed to be updated to better reflect the actions of the
method and some data no longer needed to be moved from |wants| to |has|.

The largest offender was asserts. The assert messages were not updated
when copied to new nodes.

Issue #816
Issue #997

Change-Id: I1e3e8a7c883af6a665c2223b1b2fcab330438e4e