Refine Load Graph Interruption Tests
To make it easier to read and verify, this change updates all the
interruption tests to use a single structure. The naming of the tests
was updated to make it easier to understand what they were testing,
which showed that a test case did not previously get update (it has been
updated now).

Issue #816
Issue #997

Change-Id: Ibbb70a39a5e9ea13687804448a10b24c789e03cc