Replace Destroyable.with with try-finally
The |Destroyable.with| function was a way to ensure that objects were
destroyed regardless of success or failure. With ES6 we have the
try-finally to offer this functionality.

This replaces all uses of |Destroyable.with| with try-finally with
exception of |shaka.offline.Storage|. |Storage| appears to have testing
errors if we change over to try-finally.

Issue #1516

Change-Id: Ia772b94aed047a1dd7b406fd2d716bcb64a0ca10