Isolate Move Playhead Position
In VideoWapper there is code that ensures that the current time changes
when set. To make the detail stand out more, the code was moved into
its own class. In doing so, the code that set the start time needed to
be updated. To do this, the code was refactored to better communicate
how we deferred setting the start time.

Change-Id: I98cda9d0cc0a8fed9209b361552887ab2a4425b8