Move "loading" event
The "loading" event should be fired when the user has signalled their
intent to load content. For us, this means when |Player.load| is called.

During the work on the load graph, the event was not moved, and was
being fired near the end of the load process. This change moves the
event to be fired in |Player.load|.

Issue #816
Issue #997

Change-Id: I69db702a79b70716dfd31c0e4302b5fa9f41c370