Clarify position of |adjustStartTime|
Why do we wait until after creating streaming engine to adjust the start
time of the playhead? Coming across the code, it was not clear until I
looked at the implementation of |adjustStartTime|.

Turns out that |adjustStartTime| needs the active audio and video
streams, and to get that, we need to initialize streaming engine.

To make it easier to see, this change adds comments to the call site
explaining why it is where it is.

Change-Id: Iaf1dff821b0c382f8e7e48091337cd758ded1dbc