Make Buffering Observer A Top-Level Component
Before the buffering observer was a playhead observer, but with
supporting src=, changing how we interact with the buffering observer
was made easier if it was handled as a top-level component.

This meant moving it off the playhead observer interface and create its
own timer in player.

Coming off the playehead observer interface, the buffering observer did
not need to use callbacks, which made it easier to use. This will also
allow us to use it as the single source of "is buffering" in a later
change.

Change-Id: I8cad9bfde3309de7c2b8301b90aa8c40b6e4d247