Explain why "streaming" event is fired when it is
The "streaming" event needs to be fired at a very specific time, but
that context is not visible when looking at the code where the event is
fired. To help avoid is being moved too far away from when it should be
fired, this change adds some comments to explain the nuance around its
positioning.

Change-Id: If5013ce915956cd832f66187f2558078991568af