Make Text Displayer The Authority on Text Visibility
In order to store what the text visibility should be after we load
content we have a cached value. This was causing problems with src=
content. So this changes text visibility to always be read from the text
displayer.

This allows the cached value to be a write only values. While doing this
change, some things were moved to ensure that the text visibility values
never got out a sync.

Change-Id: I15f7d00c910aaea4ddf80acd558e2819d23305f0