Rename StreamingEngine.init to StreamingEngine.start
The init method on streaming engine is responsible for initializing
streaming engine but it also starts the streaming process. In the
greater scheme of things, this is what starts playback.

To help communicate the full effect of what this one method does, this
changes its name to be "start" which helps show that it is starting a
process, not just initializing.

Change-Id: Ieceecce85917d97277782f41fd376fc6f97eaac9