Add Load Modes
Add the idea of load modes. These are low resolution state flags that
allow us to know what state we are in. These flags work with the load
graph, being set strategically during the load progress to enable and
disable public method behaviour.

In the grand-scheme of things, this solution is meant to be revisited,
but is meant to unblock the work for src=.

Issue #816
Issue #997

Change-Id: I28660d5d5ef2d746d63988b488f5d15a10ee75a4