Demo: disable unsupported manifest types
Now that our demo might be running on iOS, where only HLS will be
supported, we need to filter out unsupported manifest types in the
demo app based on the results of probeSupport().

Note that disabled options in a select element are still selectable
on some iOS versions on iphone.  This is certainly a browser bug
outside of my control, and may be obsoleted by design choices in the
upcoming replacement of our demo app.

Closes #997

Change-Id: I466e67729111bf7f3d70324e593d53e3456147a3