Fix teardown order of StreamingEngine and Playhead
In one test, we got into a situation where StreamingEngine was still
setting up a text stream when Player was destroyed.  Because Player
destroyed Playhead before StreamingEngine, StreamingEngine tried to
reference it through a callback on PlayerInterface after it was null,
but before StreamingEngine was itself destroyed.

This fixes the order of destruction in Player so that Playhead exists
longer than StreamingEngine.

b/130554111

Change-Id: I5d944e939d79de5adf6b4be3b87a254e2a4ee233