Fix PlayRateController stuck in ZERO_RATE state
PlayRateController could get stuck in a state in which playbackRate
would be stuck at 0.  It had to do with the order of modifiers
BUFFERING and ZERO_RATE.  If they occurred in that order (BUFFERING
because of underflow, followed by ZERO_RATE because of a ratechange
event), then ZERO_RATE would stick.

In some cases (~10%) we would see this as a test failure in "Player
plays while changing languages with short Periods".  In the demo app,
it could be easily reproduced by changing resolutions through the UI.

The solution is to remove ZERO_RATE, which doesn't seem to be useful.
Based on the comments in the code, this was to handle seeking in Edge,
where the browser sets playbackRate to 0 during a seek, then restores
it afterward.  If that's the case, just ignore the ratechange event
to 0, and let the browser restore it afterwards.  No need to track
that as far as I can tell.

With ZERO_RATE removed, it doesn't make sense to keep a BUFFERING
constant, so the whole set has been replaced with a boolean.

Broken in Change-Id: Id462cda2c5eb82c3713237341424b91891bd38ea
This bug did not affect any release or beta versions.

Closes #1886
b/130758100

Change-Id: I8c6455c3cdd7c9f353740425fb9337733908b2c9