Reduce starvation threshold for buffering
Commit 1284dd48, Change-Id: I8cad9bfde3309de7c2b8301b90aa8c40b6e4d247
introduced a change to the buffering logic.  Instead of going into a
buffering state when less than 0.5 seconds of content was buffered,
the thredhold changed to one half of the configured rebufferingGoal.

Now the rule is one half of the rebufferingGoal or 0.5 seconds,
whichever is smaller.  This accounts for unusually small
rebufferingGoal settings, while keeping a saner default behavior.

This change in behavior was one part of the test flake described
in internal bug report b/128923302.  The other component to the test
flake was issue #1206, which still has a separate fix.

Issue #1206
Fixes b/128923302

Change-Id: I5f133959acac6fd4ebe7fed37d2df88892e6399e