Use mdhd box over mvhd box in StreamGenerator
In our tests, we overwrite timestamps in segments to simulate streams.
In some cases, the init segment contains two different timescales:
one in the mdhd box and one in the mvhd box.  In fact, the mdhd box is
what the browser uses to interpret the tfdt box timestamps.  So the
mdhd box is what we should be reading, not the mvhd box.

This change fixes gaps in the test content generated from the multidrm
clips in our integration tests.  These were not detected before
because the tests in question never played beyond the first segment.
An unintentional change in buffering thresholds (fixed in an earlier
change) allowed us to see these gaps for the first time in our tests.

The buffering threshold change caused a buffering state in the offline
integration tests some small percentage of the time, which caused the
test to never reach its target timestamp.

This fixes the gaps by using the correct timescale to calculate the
timestamps for the tfdt box.  We knew this was an issue in theory as
far back as December 2017, but this was the first time our tests
actually failed because of it.  In most of our content, the two
timescale values match.  In the multidrm content from Axinom, they do
not.

Closes #1206

Change-Id: I863db30a16417a84f8e5df42f4d533b47fa4e7dc