Remove jsdoc links in markdown files
Jsdoc has the `@link` syntax for links, but github's markdown renderer
sees the closing curly brace as part of the link.

To format things in a way that works for both jsdoc and github, we
should specify text for the `@link` as well as the link itself.  This
separates the URL from the enclosing curly braces in a way that fixes
github's rendering of the markdown.

Change-Id: I1cb9273aae50d6089aa4d3c33a7ad5daecaa141f