Use no-cors mode for all service worker resources
Some of the resources we load from other origins must be loaded as
opaque resources that can't be read by the application.  In fact, it
turns out that there are no app resources that need to be read by the
app.  So we should load everything in no-cors mode as opaque
resources.  This fixes failures loading and caching MDL resources.

Change-Id: I25ec4228285e3149746d906e6461ac04aeedc7e1