Add a generic test waiter class
This is a centralized utility to replace the various waitFor() methods
spread out through a few utilities and suites.  This allows better
control over the formatting and type of debug info logged on failures,
as well as the timeout structure.

This utility has helped debug some tricky event timeouts in the UI
tests, and follow-up changes will put it into use throughout all test
suites.

Change-Id: I0072dcefa2c62e56e1c746f4e48cec1282934d9e