Add missing player events
If changes are deferred because of StreamingEngine state, the Player
should still fire variantchanged and textchanged events after the
changes are applied.

Also, clean up Player event firing patterns.  The texttrackvisibility
and abrstatuschanged events should be fired asynchronously, just as
other state change events are.

The onEvent_ method was one line and only called from one place, so
it was inlined and removed.

The corresponding pattern of waiting on these events in the UI tests
was also changed.

b/131909906

Change-Id: I7f7d3b25ee336dbfe79c3214854722e1955acb6a