Add support for multiple concurrent downloads.
At the moment, the storage manager does not support multiple concurrent
downloads. This modifies the demo to create a new storage instance every
time a storage operation is desired, then dispose of that instance after
the operation is complete.
This also refactors out the shaka-main-offline-changed event, which was
no longer necessary after allowing asset cards to re-load their buttons.

Issue #1432

Change-Id: I1312cf74a92f877279adb461e423a38e93bcfa0f