Fix cast receiver styling
The cast receiver should not be managing play button state or styles.
Doing this causes the play button to be off-center by interfering with
UI styles.

This also drops all outdated styles that are no longer used with the
new UI.

Reconfiguring the receiver UI controls gets rid of extra elements that
are not used in this context.

Fixes b/131254481
Fixes b/131254837

Change-Id: I25fb34a9b6a3e18fb07f3570181edbbf14a31cd3