Swap formats of front-page assets.
There is currently a bug with VTT subtitles in HLS. We plan on fixing
this bug in the future, but for v2.5, we need to switch out the
front-page assets to put our best foot forward.
Thus, this is switching out the HLS version of Angel One to the DASH
version, and the DASH version of the Shaka Player History livestream to
the HLS version. Thus, the asset with a lot of subtitles can be DASH,
while the asset with no subtitles can be HLS.
The other versions of the assets still exist; they are merely now
relegated to the search page.

google/shaka-packager#573

Change-Id: Icd2518b7caa1fe91d279e279a881c985b4acbc39