Let demo load even if storage fails to load.
Previously, if the storage instance failed to load, it would cause the
rest of the demo app to stop loading. This should suppress any errors,
so the rest of the UI can be set up at least, and just treat the failure
to load as if it was storage being unsupported.

Fixes #1925

Change-Id: Ibd969b049572fa140cec8ca789b981a9b6c659ac