Fix initial bandwidth estimate on Chrome
The Network Information API gives us an initial bandwidth estimate on
Chrome and Opera (and others, some day).  But the API changed slightly
in a way that broke our detection of it.  This restores the
functionality so that our initial bandwidth estimate is closer to what
we will converge on during playback.

Change-Id: Iaff9224c98127e6fcad1f188d73d243a45f36cd7