Defer creating contents until shown
Each of the three pages in the demo app displays asset cards.  In some
cases, these can take a long time (hundreds to thousands of
milliseconds) to create the DOM elements and load the images.  To
improve startup performance for the app, defer creating the contents
of any given page until the page is shown.

This improves the demo's performance score in Lighthouse from 54 to
82, the "First Contentful Paint" score by 43%, and the "First
Meaningful Paint" score by 10%.

This incidentally fixes a race condition between the asset setup and
calculation of the platform's capabilities, so that the buttons are
always in the correct state on startup.

Change-Id: I73a70dec89e13cc16566dde7ef8659b3e1c64b17