Move download buttons to corner.
Instead of the download button being a button called "Download" or
"Delete", this moves it to the upper-right corner and made it into an
MDL icon instead of text. That way, there's no confusion between
removing a custom asset from storage and removing a custom asset from
the assets list.

Change-Id: I74d426bc9e9030598b741afc58a3ba7d6cec4c4a