Clean up EBML size constants
Instead of using an array of Uint8Arrays, which could trigger a
load-time failure on very old browsers, store an array of number
arrays.  This is equivalent for the purpose of comparison, and it
cannot fail at load-time.

This cleanup seems to fix minor test flake in the EBML parser tests
that was observed with upgraded node modules.

Issue #1694

Change-Id: I87b2307bceddb3a21dfc81d4fe9828afd9508617