Fix bug with text displayer switching.
The new support for replacing the text displayer mid-playback had an
uncaught problem. Specifically, loading a new asset while an old asset
was playing caused an error.
This was because it was trying to reload the existing text stream when
there was no text stream.

This changes the reloadTextStream method on streaming engine to do
nothing if there is no text stream, fixing that problem.

Change-Id: I53f25fb14a54670697ad014bc6e6f72f6cc28577