Made demo properly re-connect to cast sessions.
Note that this does not fully fix the problem; issue #1948 means
that if the ui is configured more than once, we cannot cast at all.
Also, the cast receiver does not currently work; once a v2.5.1 version
is released that will no longer be an issue, but the v2.5.0 version
currently uploaded is missing goog.require calls.
I have manually verified that once those issues are fixed, though,
this will allow for the player to re-connect.

Issue #1948

Change-Id: I26201cc696a61906fac3c320f9ac5e4b61fe1920