Update the extern generator for ES6 data classes
For classes which only contain a bunch of public, exported members set
in the constructor, the extern generator goes through some extra steps
to create externs for those.  This updates the generator to handle
these as they are converted to ES6.

Note that we currently have only three such classes:
 - shaka.util.Error
 - shaka.text.Cue
 - shaka.text.CueRegion

Issue #1157

Change-Id: I8041998b2bdca0e6f44dcc6f897eb3c660b3424f