Removed requirement for JSDoc for constructors
Previously, generateExterns required all methods to have JSDoc.
This made sense most of the time, but for an ES6 constructor not as
much. What the constructor does is kind of self-explanatory, and a JSDoc
along the lines of "/** Creates a Foo object. */" is a waste of space.

This changes generateExterns to not require a comment for such
constructors.

Change-Id: I12ae019934de9c020340baf0ca25f940d58d7d2f