Fix coverage reports
This changes from using Karma's "coverage" preprocessor to using a
Babel plugin called "istanbul" which adds the coverage instrumentation
to the code.  Because it happens in Babel, all the code
transformations happen in one place, with consistent source maps.
Backtraces in error messages will show the original line numbers, and
code coverage reports will refer to the original line numbers, as
well.

Closes #1693

Change-Id: I706cf21c109f3ec7fa30c6899ca9872c0f8be9d1